+++ app/src/main/java/com/example/cardgame/GameplayFragment.kt
@@ -10,6 +10,9 @@ import android.view.ViewGroup
 import android.widget.Button
 import android.widget.ImageView
 import android.widget.TextView
+import com.example.cardgame.models.CardInfo
+import com.example.cardgame.models.ResponseForNewCard
+import com.example.cardgame.models.ResponseNewDeck
 import com.example.cardgame.utils.Consts
 import com.google.gson.Gson
 import com.google.gson.annotations.SerializedName
@@ -164,64 +167,6 @@ class GameplayFragment : Fragment() {
 
         Log.d(TAG, "onPause")
     }
-
-
-    //toDo Put these classes in a separate folder
-    // MODELS
-    data class ResponseNewDeck(
-        @field:SerializedName("success")
-        val success: Boolean,
-
-        @field:SerializedName("deck_id")
-        val deckId: String,
-
-        @field:SerializedName("shuffled")
-        val shuffled: Boolean,
-
-        @field:SerializedName("remaining")
-        val remaining: Int
-    )
-
-    data class ResponseForNewCard(
-        @field:SerializedName("cards")
-        val cards: Array<CardInfo>,
-
-        @field:SerializedName("remaining")
-        val remaining: String,
-
-        @field:SerializedName("deck_id")
-        val deckId: String,
-
-        @field:SerializedName("success")
-        val success: Boolean
-    )
-
-    data class CardInfo(
-        @field:SerializedName("images")
-        val images: ImageData,
-
-        @field:SerializedName("image")
-        val image: String,
-
-        @field:SerializedName("value")
-        val value: String,
-
-        @field:SerializedName("code")
-        val code: String,
-
-        @field:SerializedName("suit")
-        val suit: String
-    )
-
-    data class ImageData(
-
-        @field:SerializedName("svg")
-        val svg: String,
-
-        @field:SerializedName("png")
-        val png: String
-    )
-
 
     private fun getCardObservable() = Observable.fromCallable {
